refactor(charging-station): extract charging-profile helpers to HelpersChargingProfile (#1936 f-2)
Split Helpers.ts (Phase 2 of 5). Moves 14 charging-profile symbols into
sibling HelpersChargingProfile.ts and re-exports the public ones from
Helpers.ts so the barrel import path is preserved. Zero behavior change:
same signatures, same code, same log content. The moduleName local in
the new file switches log prefixes from 'Helpers.<fn>' to
'HelpersChargingProfile.<fn>', reflecting the actual source location.

Moved (private inside new file):
- getChargingStationChargingProfiles, buildChargingProfilesLimit,
  ChargingProfilesLimit interface, getChargingProfileId,
  getChargingProfilesLimit, canProceedRecurringChargingProfile,
  prepareRecurringChargingProfile, checkRecurringChargingProfileDuration

Moved (exported from new file):
- getChargingStationChargingProfilesLimit, getConnectorChargingProfiles,
  getConnectorChargingProfilesLimit, prepareChargingProfileKind,
  canProceedChargingProfile (all re-exported from Helpers.ts)
- getSingleChargingSchedule (exported for Phase 3 direct import; not
  re-exported from Helpers.ts to keep the public API unchanged)

Follows the barrel pattern established by HelpersReservation.ts in
Phase 1 (PR #1946).
